requirements.txt
Django==2.1.5
djangorestframework==3.9.1
django-extensions==2.1.4
coreapi==2.3.3
ipython==7.2.0